feat: use sveltekit
