+++ package.json
@@ -6,24 +6,21 @@
     "dev": "vite",
     "build": "vite build",
     "preview": "vite preview",
-    "check": "svelte-check --tsconfig ./tsconfig.json",
+    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
     "test": "vitest"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^2.5.3",
     "@testing-library/svelte": "^4.0.3",
-    "@tsconfig/svelte": "^5.0.2",
     "svelte": "^4.2.8",
     "svelte-check": "^3.6.2",
     "svelte-preprocess": "^5.1.1",
-    "svelte-spa-router": "^3.3.0",
     "typescript": "^5.3.3",
     "vite": "^4.5.1",
     "vitest": "^0.34.6"
   },
   "dependencies": {
+    "@sveltejs/kit": "^1.30.3",
     "dayjs": "^1.11.10",
-    "dotenv": "^16.3.1",
     "firebase": "^10.7.1"
   },
   "engines": {
