+++ src/app.html
@@ -2,12 +2,10 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width,initial-scale=1" />
-
+		<link rel="icon" href="%sveltekit.assets%/favicon.ico" />
+		<meta name="viewport" content="width=device-width" />
+		%sveltekit.head%
     <title>Inishie</title>
-
-    <link rel="icon" type="image/png" href="/favicon.ico" />
-    <link rel="stylesheet" href="/global.css" />
     <!-- Global site tag (gtag.js) - Google Analytics -->
     <script
       async
@@ -39,7 +37,6 @@
   </head>
 
   <body>
-    <div id="app"></div>
-    <script type="module" src="/src/main.ts"></script>
+		<div style="display: contents">%sveltekit.body%</div>
   </body>
 </html>
