--- src/components/BackToHome.spec.ts
@@ -1,16 +1,0 @@
-// import { render } from "@testing-library/svelte";
-import { describe, test } from "vitest";
-// import Component from './BackToHome.svelte';
-
-// todo テストのみだと validate でエラーが出るため空 export を追加
-export {};
-
-describe("BackToHome", () => {
-  // NOTE: svelte-spa-router が解決できずエラーになるため、一旦コメントアウトしておく
-  // it('戻るボタンのリンクに /home が設定されている', () => {
-  //   const { getByText } = render(Component);
-
-  //   expect(getByText('戻る').getAttribute('href')).toBe('/home');
-  // });
-  test.todo("戻るボタンのリンクに /home が設定されている");
-});
