+++ src/lib/components/BackToHome.svelte
@@ -1,0 +1,9 @@
+<div class="back-link">
+  <a href={"/home"}>戻る</a>
+</div>
+
+<style>
+  .back-link {
+    margin: 5px auto;
+  }
+</style>
