+++ src/lib/components/Description.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import Counter from "@/components/Counter.svelte";
+  import Counter from "$lib/components/Counter.svelte";
 
   const names = ["太郎", "次郎", "三郎"];
   const name = names[Math.floor(Math.random() * names.length)];
