+++ src/lib/components/Entrance.spec.ts
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/svelte";
 import { describe, expect, test } from "vitest";
-import Component from "@/components/Entrance.svelte";
+import Component from "$lib/components/Entrance.svelte";
 
 describe("Entrance", () => {
   test("10件の入口が表示される", () => {
