+++ src/routes/+page.svelte
@@ -1,14 +1,14 @@
 <script lang="ts">
   import { onMount } from "svelte";
-  import { push } from "svelte-spa-router";
-  import AppDescription from "@/components/Description.svelte";
-  import AppFooter from "@/components/Footer.svelte";
-  import AppEntrance from "@/components/Entrance.svelte";
-  import AccessLogRepository from "@/repositories/access_log";
+  import { goto } from '$app/navigation';
+  import AppDescription from "$lib/components/Description.svelte";
+  import AppFooter from "$lib/components/Footer.svelte";
+  import AppEntrance from "$lib/components/Entrance.svelte";
+  import AccessLogRepository from "$lib/repositories/access_log";
 
   const repo = new AccessLogRepository();
 
-  const moveToHome = () => push("/home");
+  const moveToHome = () => goto("/home");
 
   if (!(import.meta as any).env.DEV) {
     onMount(async () => {
