+++ svelte.config.js
@@ -1,5 +1,13 @@
-import { vitePreprocess } from "@sveltejs/vite-plugin-svelte";
+import { vitePreprocess } from "@sveltejs/kit/vite";
 
-export default {
+/** @type {import('@sveltejs/kit').Config} */
+const config = {
   preprocess: vitePreprocess(),
+	kit: {
+		alias: {
+			$routes: './src/routes'
+		},
+  }
 };
+
+export default config;
